+++ SpaceArcaders/src/net/awhipple/spacearcaders/utils/GameMath.java
@@ -9,11 +9,11 @@ package net.awhipple.spacearcaders.utils;
  * @author Aaron
  */
 public class GameMath {
-    public static double pointsToRad(float x1, float y1, float x2, float y2) {
+    public static double pointsToRad(double x1, double y1, double x2, double y2) {
         double pi = Math.PI;
         if(x1==x2) {
-            if(y1>y2) return (3*pi)/2f;
-            else return pi/2f;
+            if(y1>y2) return (3*pi)/2d;
+            else return pi/2d;
         }
         double angle = Math.atan((y2-y1)/(x2-x1));
         if(x2<x1) { return pi+angle; }
@@ -22,7 +22,7 @@ public class GameMath {
     }
     
     //returns -1 if p1 < p2, 0 if p1==p2 and 1 if p1 > p2
-    public static int whichSide(float p1, float p2) {
+    public static int whichSide(double p1, double p2) {
         if(p1 < p2) return -1;
         else if (p1 > p2) return 1;
         else return 0;
